# Perf: Adding torch.compile + static cache, ~3x speed up