Did:
- Add Static KVCache from gpt-fast
- Adjust logic to use new KV Cache
- Fix code for torch.compile compatible

Result on A100 for 1000 tokens:
Overall:
**~x2.6 faster for fp32
~x3.2 faster for fp16**


```
Branch main: Using device: cuda
<torch.utils.benchmark.utils.common.Measurement object at 0x7f2b05192c50>
generate_tokens(tokens, image)
setup: from __main__ import generate_tokens
  23.38 s
  1 measurement, 10 runs , 24 threads   
Using device: cuda with dtype: torch.float32, torch.compile: False
<torch.utils.benchmark.utils.common.Measurement object at 0x7f8015c3df90>
generate_tokens(tokens, image)
setup: from __main__ import generate_tokens
  26.81 s
  1 measurement, 10 runs , 24 threads

Using device: cuda with dtype: torch.float32,  torch.compile: True
setup: from __main__ import generate_tokens
  9.97 s
  1 measurement, 10 runs , 24 threads


Using device: cuda with dtype: torch.float16, torch.compile: False
<torch.utils.benchmark.utils.common.Measurement object at 0x7f7c70058310>
generate_tokens(tokens, image)
setup: from __main__ import generate_tokens
  26.65 s
  1 measurement, 10 runs , 24 threads


Using device: cuda with dtype: torch.float16, torch.compile: True
<torch.utils.benchmark.utils.common.Measurement object at 0x7fe84850edd0>
generate_tokens(tokens, image)
setup: from __main__ import generate_tokens
  7.84 s
  1 measurement, 10 runs , 24 threads
```


TODO:
- Adjust code for compatible with previous KVCache
- Check training compatible