@tsdocode 2025-05-28
This pr is WIP, but above code is testable

@andimarafioti @lusxvr  If you guy have a moment, please help me give it a shot!