@tsdocode 2025-06-03
> I was looking at it, is this 3x faster that the current kv-cache implementation? How much of that is the torch compile and how mache the kv cache implementation?

3x faster comefrom the combination of static kvcache + torch.compile:
- torch.compile require static shape in inference, unless it will recompile everystep => why we need implement static kvcache
- without torch compile:
  - dynamic kvcache: speed is nearly the same as previous
  - static shape: speed is a little lower on longer sequence length