@tsdocode 2025-06-08
Update new benchmark result with torch.compile + static cache + fp16 after separate prefill and decode method in LLM, it is now ~5x faster:
```
(f5) ➜  nanoVLM git:(perf/torch-compile) ✗ python benchmark-inference.py
Using device: cuda with dtype: torch.float16
<torch.utils.benchmark.utils.common.Measurement object at 0x7ff7ae582a90>
generate_tokens(tokens, image)
setup: from __main__ import generate_tokens
  5.52 s
  1 measurement, 10 runs , 24 threads
```

@andimarafioti please help me advise the next step!